requirements.txt
streamlit>=1.24
pandas>=1.5
pillow>=10
openpyxl>=3.0
matplotlib>=3.5
pyperclip>=1.8.2
